@@ -2,8 +2,7 @@ package b.softuni.surfApp.model.entity;
 
 
 import b.softuni.surfApp.model.enums.EquipmentTypeEnum;
-
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 @Table(name = "equipment_types")
