@@ -1,8 +1,7 @@
 package b.softuni.surfApp.model.entity;
 
-import b.softuni.surfApp.model.enums.EquipmentTypeEnum;
+import jakarta.persistence.*;
 
-import javax.persistence.*;
 import java.util.Objects;
 
 @Entity
