@@ -2,6 +2,7 @@ package b.softuni.surfApp.web;
 
 import b.softuni.surfApp.model.binding.UserRegisterBindingModel;
 import b.softuni.surfApp.service.UserService;
+import jakarta.validation.Valid;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -10,8 +11,6 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
-
-import javax.validation.Valid;
 
 @Controller
 public class UserRegistrationController {
